repository_api/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
fastapi==0.75.2 ; python_version >= "3.9" and python_full_version < "3.13"
uvicorn==0.17.6 ; python_version >= "3.9" and python_full_version < "3.13"
transformers==4.49.0 ; python_version >= "3.9" and python_full_version < "3.13"
qdrant-client==1.12.0 ; python_version >= "3.9" and python_full_version < "3.13"
python-dotenv==0.21.1 ; python_version >= "3.9" and python_full_version < "3.13"
sentence-transformers==2.7.0 ; python_version >= "3.9" and python_full_version < "3.13"
tqdm>=4.66.0 ; python_version >= "3.9" and python_full_version < "3.13"
numpy==1.26.4 ; python_version >= "3.9" and python_full_version < "3.13"
tree-sitter-language-pack==0.10.0 ; python_version >= "3.9" and python_full_version < "3.13"
pathspec==0.12.1 ; python_version >= "3.9" and python_full_version < "3.13"
fastembed-gpu>=0.8.0 ; python_version >= "3.9" and python_full_version < "3.13"